added some java doc

+++ JavaMacro/src/macro/instruction/ScriptContainer.java
@@ -1,19 +1,17 @@
 package macro.instruction;
 
-import com.github.kwhat.jnativehook.keyboard.NativeKeyEvent;
-import macro.Main;
-
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ScriptContainer {
 
-	private final Map<Integer, InstructionSet> instructionSetMap = new HashMap<>();
+	private final Map<Short, InstructionSet> instructionSetMap = new HashMap<>();
 
 	public void insert(final InstructionSet instructionSet) {
 		instructionSetMap.put(instructionSet.key, instructionSet);
 	}
 
-	public final Map<Integer, InstructionSet> getInstructionSetMap() {
+	public final Map<Short, InstructionSet> getInstructionSetMap() {
 		return this.instructionSetMap;
 	}
 
@@ -23,7 +21,7 @@ public class ScriptContainer {
 
 	//resets all locks back to false
 	public void clearLocks() {
-		for (Map.Entry<Integer, InstructionSet> entry : instructionSetMap.entrySet()) {
+		for (Map.Entry<Short, InstructionSet> entry : instructionSetMap.entrySet()) {
 			InstructionSet set = entry.getValue();
 			set.bFlags &= ~0x08;
 		}
